ElementDB
- move to app package
- concurrentElementDB
